Moves WaitTillProvisioned* methods and add basic test
