+++ datacenter.go
@@ -1,8 +1,6 @@
 package profitbricks
 
 import (
-	"context"
-	"errors"
 	"net/http"
 	"strconv"
 	"time"
@@ -93,43 +91,4 @@ func (c *Client) DeleteDatacenter(dcid string) (*http.Header, error) {
 	url := dcPath(dcid) + `?depth=` + c.client.depth + `&pretty=` + strconv.FormatBool(c.client.pretty)
 	ret := &http.Header{}
 	return ret, c.client.Delete(url, ret, http.StatusAccepted)
-}
-
-// WaitTillProvisionedOrCanceled waits for a request to be completed.
-// It returns an error if the request status could not be fetched, the request
-// failed or the given context is canceled.
-func (c *Client) WaitTillProvisionedOrCanceled(ctx context.Context, path string) error {
-	ticker := time.NewTicker(1 * time.Second)
-	defer ticker.Stop()
-	for {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		case <-ticker.C:
-			request, err := c.GetRequestStatus(path)
-			if err != nil {
-				return err
-			}
-			switch request.Metadata.Status {
-			case "DONE":
-				return nil
-			case "FAILED":
-				return errors.New("request failed")
-			}
-		}
-	}
-}
-
-// WaitTillProvisioned waits for a request to be completed.
-// It returns an error if the request status could not be fetched, the request
-// failed or a timeout of 2.5 minutes is exceeded.
-func (c *Client) WaitTillProvisioned(path string) (err error) {
-	ctx, cancel := context.WithTimeout(context.TODO(), 150*time.Second)
-	defer cancel()
-	if err = c.WaitTillProvisionedOrCanceled(ctx, path); err != nil {
-		if err == context.DeadlineExceeded {
-			return errors.New("timeout expired while waiting for request to complete")
-		}
-	}
-	return
 }
